pyproject.toml
[project]
name = "nutrient-analysis-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "rapidfuzz>=3.14.3",
    "uvicorn>=0.42.0",
]
